+++ src/DependencyInjection/DoctrinePHPCRExtension.php
@@ -126,7 +126,6 @@ final class DoctrinePHPCRExtension extends AbstractDoctrineExtension
 
             $type = $session['backend']['type'];
             switch ($type) {
-                case 'prismic':
                 case 'doctrinedbal':
                 case 'jackrabbit':
                     if (empty($loaded['jackalope'])) {
@@ -209,10 +208,6 @@ final class DoctrinePHPCRExtension extends AbstractDoctrineExtension
                         $backendParameters['jackalope.data_caches'][$key] = new Reference($cache);
                     }
                 }
-
-                break;
-            case 'prismic':
-                $backendParameters['jackalope.prismic_uri'] = $session['backend']['url'];
 
                 break;
             case 'jackrabbit':